Fix dependencies to be platformed and fix newly enabled tests
