+++ core/jvm/src/test/scala/specs/AsynchronousLoggingSpec.scala
@@ -1,4 +1,4 @@
-package spec
+package specs
 
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
