+++ core/shared/src/test/scala/specs/MDCSpec.scala
@@ -37,37 +37,5 @@ class MDCSpec extends AnyWordSpec with Matchers {
         ))
       }
     }
-    "verify concurrency access to implicits" in {
-      implicit val mdc: MDC = MDC.instance
-      logger.info("Zero")
-      mdc.context("test" -> "testing") {
-        logger.info("One")
-        mdc.context("test" -> "testing2") {
-          logger.info("Two")
-          var finished = false
-          new Thread {
-            override def run(): Unit = {
-              logger.info("Three")
-              finished = true
-            }
-          }.start()
-          while (!finished) {
-            Thread.sleep(10)
-          }
-        }
-        logger.info("Four")
-      }
-      logger.info("Five")
-      writer.consume { list =>
-        list.map(r => r.messages.head.logOutput.plainText -> r.data.get("test").map(_())) should be(List(
-          "Five" -> None,
-          "Four" -> Some("testing"),
-          "Three" -> Some("testing2"),
-          "Two" -> Some("testing2"),
-          "One" -> Some("testing"),
-          "Zero" -> None
-        ))
-      }
-    }
   }
 }
