+++ fileModule/jvm/src/test/scala/spec/FileLoggingSpec.scala
